Project Finished; Ready for Review
In a quick effort, I already have debugged the project. I'm not sure if this was the intended solution, but I simply replaced the "&&" symbols with "||" in the "testFormCompleteness" function and that solved all issues with the recommendations flawlessly. With that said, I uploaded the completed .zip file to Google classroom meaning that I'm sure this project is now whole and complete. (Guess I'll finish up on those review questions now)